Added functionality to send various types of data between users.
This commit adds a new functional to the codebase that allows users to send various types of data (including voice, video, text, audio, and photo) between each other. Previously, there were separate methods for sending each type of data, but in the future we created a single method for each type of data.