Added chat functionality improvements and centralization of messages.
Refactored code to fix the issue of incorrect connection between users in the chat. Implemented a feature to change reputation after the chat. Moved all string messages to the MessagesText class to centralize and manage all the text in one place. Created a MessageBuilder and KeyboardBuilder classes to simplify the creation of messages and keyboards. Added a new method, saveAll, to the BaseService class to facilitate saving all data. Added the previousChatId field to the Chat class, which helps implement the reputation functionality. Also added a new sendMessage method to the MessageSender class for better message sending functionality.