Improved message sending, reputation, reporting, blocking, and logging.
Improved logic for sending messages between users by moving it to the MessageSender class. Implemented functionality to change reputation and set reports: when a user reports another user, the reported user gets blocked. Also added a list of blocked users for each user. Updated the list of reputation items. Added a logger to the MessageSender file. Added a new string to the MessagesText file.