Improved Messaging, User Model, and Reports
Fixed an issue with exceptions occurring when a message does not contain any text. Added a command to retrieve account information. Added values to the Report enum. Updated the User model. Added new items for reputationItems. Implemented null checks for MessageSender methods. Updated the start message. Improved the connection between two users and fixed the functionality of setting reports for a user.